requirements.txt
streamlit
openai
pandas
scikit-learn
google-search-results
pydrive2
google-auth
oauth2client